updated code to implement INotification
