+++ Application/Command/CreateOrder/CreateOrderCommand.cs
@@ -3,7 +3,7 @@ using MediatrWithPublish.Domain;
 
 namespace MediatrWithPublish.Application.Command.CreateOrder;
 
-public class CreateOrderCommand :  IRequest<Order>
+public class CreateOrderCommand : IRequest<Order>
 {
     public string CustomerName { get; set; }
     public decimal TotalAmount { get; set; }
