+++ Application/Command/CreateOrder/CreateOrderCommandHandler.cs
@@ -3,19 +3,18 @@ using MediatrWithPublish.Domain;
 
 namespace MediatrWithPublish.Application.Command.CreateOrder;
 
-public class CreateOrderCommandHandler: IRequestHandler<CreateOrderCommand, Order>
+public class CreateOrderCommandHandler : IRequestHandler<CreateOrderCommand, Order>
 {
     private readonly IMediator _mediator;
-    
+
     public CreateOrderCommandHandler(IMediator mediator)
     {
         _mediator = mediator;
     }
-    
-    
+
+
     public async Task<Order> Handle(CreateOrderCommand request, CancellationToken cancellationToken)
     {
-        // Validate the request
         // Create the order in the database
         var order = new Order
         {
@@ -25,9 +24,9 @@ public class CreateOrderCommandHandler: IRequestHandler<CreateOrderCommand, Orde
         };
 
         // Publish a notification that the order has been created
-        await _mediator.Publish(new OrderCreatedNotification() { Order  = order }, cancellationToken);
+        await _mediator.Publish(new OrderCreatedNotification() { Order = order }, cancellationToken);
 
         return order;
-        
+
     }
 }
