--- Application/Command/CreateOrder/OrderCreatedNotification.cs
@@ -1,9 +1,0 @@
-﻿using MediatR;
-using MediatrWithPublish.Domain;
-
-namespace MediatrWithPublish.Application.Command.CreateOrder;
-
-public class OrderCreatedNotification : INotification
-{
-    public Order Order { get; set; }
-}
