+++ Domain/OrderCreatedNotification.cs
@@ -1,0 +1,9 @@
+﻿using MediatR;
+
+namespace MediatrWithPublish.Domain
+{
+    public class OrderCreatedNotification : INotification
+    {
+        public required Order Order { get; set; }
+    }
+}
