--- global.json
@@ -1,7 +1,0 @@
-{
-  "sdk": {
-    "version": "6.0.0",
-    "rollForward": "latestMinor",
-    "allowPrerelease": false
-  }
-}
